feat: 0-100000 indexed and packages updated with checksum

+++ automate.py
@@ -4,14 +4,14 @@ failed = []
 
 for file in os.listdir('packages'):
     name = file.replace('.json', '')
-    
-    pid = os.system(rf'target\release\registry.exe {name}')
-    if pid == 1:
-        failed.append(pid)
-    
-    try:
-        os.remove(file)
-    except:
-        pass
+    if len(name) != 128:
+        pid = os.system(rf'target\release\registry.exe {name}')
+        if pid == 1:
+            failed.append(pid)
+        
+        try:
+            os.remove(file)
+        except:
+            pass
 
 print(failed)
